image-server/requirements.txt
tensorflow
tensorflow_hub
opencv-python
pytesseract
Pillow